Pug build and test on multiple platforms  (#8)
The main aim of this PR is to add expect based tests for pug on each platform: linux, mac and windows.

Python's pexpect library is used for linux/mac and wexpect for windows. A bug was found in wexpect so my private Github fork of it is used.

The linux tests are run on each push to any branch.
The multiplatform tests are run when PRs are created and for any push to main.